# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Book Favorites API
  description: A REST API for managing book collections, user favorites, reading lists, and reviews
  version: 1.0.0
  contact:
    name: API Support
    email: support@bookfavorites.com
servers:
  - url: http://localhost:4000/api/v1
    description: Development server
  - url: http://localhost:4000/api
    description: Development server (legacy, redirects to v1)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login
  schemas:
...
    Book:
      type: object
      properties:
        id:
          type: string
          example: "1"
        title:
          type: string
          example: "To Kill a Mockingbird"
        author:
          type: string
          example: "Harper Lee"
        genre:
          type: string
          example: "Fiction"
...
    BookStats:
      type: object
      properties:
        totalBooks:
          type: integer
          example: 50
        favoritedBooks:
          type: integer
          example: 25
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "john_doe"
        password:
          type: string
          format: password
          example: "password123"
    RegisterRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 50
          example: "john_doe"
        password:
          type: string
          minLength: 1
          maxLength: 72
          format: password
          example: "password123"
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
...
    AddFavoriteRequest:
      type: object
      required:
        - bookId
      properties:
        bookId:
          type: string
          example: "1"
...
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid input data"
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 50
        pages:
          type: integer
          example: 5
    Links:
      type: object
      properties:
        self:
          type: string
          example: "/api/v1/books?page=1&limit=10"
        next:
          type: string
          nullable: true
          example: "/api/v1/books?page=2&limit=10"
        prev:
          type: string
          nullable: true
          example: ""
    PaginatedBooksResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Book'
        pagination:
          $ref: '#/components/schemas/Pagination'
        links:
          $ref: '#/components/schemas/Links'
    SearchResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Book'
        pagination:
          $ref: '#/components/schemas/Pagination'
        query:
          type: string
          example: "mockingbird"
paths:
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validation_error:
                  value:
                    error:
                      code: "VALIDATION_ERROR"
                      message: "Invalid username or password format."
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                user_exists:
                  value:
                    error:
                      code: "USER_EXISTS"
                      message: "User already exists."
        '429':
          description: Rate limit exceeded
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              description: Request limit per window
            X-RateLimit-Remaining:
              schema:
                type: integer
              description: Remaining requests in current window
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      summary: Authenticate user and get JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials or missing input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_input:
                  value:
                    error:
                      code: "INVALID_INPUT"
                      message: "Username and password are required."
                invalid_credentials:
                  value:
                    error:
                      code: "INVALID_CREDENTIALS"
                      message: "Invalid username or password."
        '429':
          description: Rate limit exceeded
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /books:
    get:
      summary: Get all books with pagination, filtering, and sorting
      tags:
        - Books
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of items per page
        - name: sort
          in: query
          schema:
            type: string
            enum: [title, author, genre]
            default: title
          description: Sort field
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort order
        - name: genre
          in: query
          schema:
            type: string
          description: Filter by genre (case-insensitive contains)
        - name: author
          in: query
          schema:
            type: string
          description: Filter by author (case-insensitive contains)
      responses:
        '200':
          description: Paginated list of books
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBooksResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /books/search:
    get:
      summary: Search books by title
      tags:
        - Books
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query for title (case-insensitive contains)
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Search results with pagination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Missing or invalid search query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_query:
                  value:
                    error:
                      code: "VALIDATION_ERROR"
                      message: "Query parameter \"q\" is required"
  /books/stats:
    get:
      summary: Get book statistics
      tags:
        - Books
      responses:
        '200':
          description: Book statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookStats'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /favorites:
    get:
      summary: Get user's favorite books
      tags:
        - Favorites
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of user's favorite books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '401':
          description: Authentication required
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                user_not_found:
                  value:
                    error:
                      code: "USER_NOT_FOUND"
                      message: "User not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Add book to user's favorites
      tags:
        - Favorites
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddFavoriteRequest'
      responses:
        '201':
          description: Book added to favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          description: Invalid book ID or book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validation_error:
                  value:
                    error:
                      code: "VALIDATION_ERROR"
                      message: "Book ID is required"
                book_not_found:
                  value:
                    error:
                      code: "BOOK_NOT_FOUND"
                      message: "Book not found"
        '401':
          description: Authentication required
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /favorites/{bookId}:
    delete:
      summary: Remove book from user's favorites
      tags:
        - Favorites
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
          description: ID of the book to remove
      responses:
        '204':
          description: Book removed from favorites
        '401':
          description: Authentication required
        '404':
          description: User not found or book not in favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                user_not_found:
                  value:
                    error:
                      code: "USER_NOT_FOUND"
                      message: "User not found"
                book_not_in_favorites:
                  value:
                    error:
                      code: "BOOK_NOT_IN_FAVORITES"
                      message: "Book not in favorites"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...